Moved NocaseDict and internal utility functions into their own submodules
Details:
- For general cleanup, moved class `NocaseDict` into its own submodule
  `_nocasedict.py`. Its unit test cases already had been in a separate
  sub module (`test_nocasedict.py`).
- In order to untangle existing circular import dependencies, as well as
  new ones that were created by the move of NocaseDict, moved the following
  utility functions into a new submodule `_utils.py`:
  - `_stacklevel_above_module()`
  - `_ensure_unicode()`
  - `_ensure_bytes()`
  - `_ensure_bool()`
  - `_hash_name()`
  - `_hash_item()`
  - `_hash_dict()`
- Adjusted any code that imports the moved symbols, to now import them from
  their new submodules. As part of that, several runtime-imports could
  be moved back to become module-level imports.
- Adjusted the documentatzion for the new module structure. This did not
  result in any changes to the generated documentation.
