+++ pywbem/_server.py
@@ -100,7 +100,8 @@ from .cim_constants import CIM_ERR_INVALID_NAMESPACE, CIM_ERR_INVALID_CLASS, \
     CIM_ERR_METHOD_NOT_FOUND, CIM_ERR_METHOD_NOT_AVAILABLE, \
     CIM_ERR_NOT_SUPPORTED, CIM_ERR_NOT_FOUND
 from .exceptions import CIMError
-from .cim_obj import CIMInstanceName, NocaseDict
+from ._nocasedict import NocaseDict
+from .cim_obj import CIMInstanceName
 from .cim_operations import WBEMConnection
 from ._valuemapping import ValueMapping
