+++ testsuite/test_cim_obj.py
@@ -35,7 +35,7 @@ from pywbem import CIMInstance, CIMInstanceName, CIMClass, CIMClassName, \
     CIMProperty, CIMMethod, CIMParameter, CIMQualifier, \
     CIMQualifierDeclaration, Uint8, Uint16, Uint32, \
     Uint64, Sint8, Sint16, Sint32, Sint64, Real32, Real64, CIMDateTime
-from pywbem.cim_obj import NocaseDict
+from pywbem._nocasedict import NocaseDict
 from pywbem.cim_types import _Longint
 
 try:
